XMLHttpRequest/abort-after-receive.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: abort() after successful receive</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-abort()-method" />
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test();

      test.step(function() {
        var client = new XMLHttpRequest(),
            control_state = 0;

        client.onreadystatechange = function() {
          test.step(function() {
            if (client.readyState == 4) {
              control_state = 1;

              // abort should not cause the "abort" event to fire
              client.abort();

              assert_equals(control_state, 1);

              test.done();
            }
          })
        };

        client.onabort = function () {
            // this should not fire!

            control_state = 2;
        };

        client.open("GET", "resources/well-formed.xml", true);
        client.send(null);
      });
    </script>
  </body>
</html>
